pyproject.toml
[tool.poetry]
name = "py-ai-chatbot-study"
version = "0.1.0"
description = ""
authors = ["therealmcfly"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
langchain = "^0.2.13"
langchain-openai = "^0.1.21"
langchain-community = "^0.2.12"
chromadb = "^0.5.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
